Refactor fromBob and fromJson into embeddedDisplay.
