+++ public/json/beamline.json
@@ -29,9 +29,10 @@
               "text": "Synoptic"
             },
             {
-              "type": "widgetFromJSON",
+              "type": "embeddedDisplay",
               "position": "relative",
-              "file": "synoptic.json"
+              "file": "synoptic.json",
+              "filetype": "json"
             }
           ]
         },
