+++ public/json/digitelmpcionp.json
@@ -3,7 +3,8 @@
   "position": "relative",
   "children": [
     {
-      "type": "widgetFromBob",
+      "type": "embeddedDisplay",
+      "filetype": "opi",
       "position": "relative",
       "margin": "10px",
       "file": "digitelMpcIonp.opi",
