+++ public/json/motorSummary.json
@@ -3,28 +3,31 @@
   "position": "relative",
   "children": [
     {
-      "type": "widgetFromJSON",
+      "type": "embeddedDisplay",
       "position": "relative",
       "margin": "10px",
       "file": "motor.json",
+      "filetype": "json",
       "macroMap": {
         "motor": "1"
       }
     },
     {
-      "type": "widgetFromJSON",
+      "type": "embeddedDisplay",
       "position": "relative",
       "margin": "10px",
       "file": "motor.json",
+      "filetype": "json",
       "macroMap": {
         "motor": "2"
       }
     },
     {
-      "type": "widgetFromJSON",
+      "type": "embeddedDisplay",
       "position": "relative",
       "margin": "10px",
       "file": "motor.json",
+      "filetype": "json",
       "macroMap": {
         "motor": "3"
       }
