+++ src/ui/widgets/EmbeddedDisplay/embeddedDisplay.test.tsx
@@ -1,29 +1,32 @@
 import React from "react";
-import { WidgetFromJson } from "./fromJson";
+import { EmbeddedDisplay } from "./embeddedDisplay";
 import { shallow } from "enzyme";
 
 import { Display } from "../Display/display";
 import { Label } from "../Label/label";
 import { DEFAULT_BASE_URL } from "../../../baseUrl";
 
-beforeEach((): void => {
-  const useEffect = jest.spyOn(React, "useEffect");
-  const mockUseEffect = (): void => {
-    useEffect.mockImplementationOnce((f): any => f());
-  };
+const useEffect = jest.spyOn(React, "useEffect");
+const mockUseEffect = (): void => {
+  useEffect.mockImplementationOnce((f): any => f());
+};
 
+beforeEach((): void => {
   mockUseEffect();
 });
 
-describe("<WidgetFromJson>", (): void => {
+describe("<EmbeddedDisplay>", (): void => {
   it.each<any>([
-    ["TestFile", `${DEFAULT_BASE_URL}/json/TestFile`],
-    ["https://a.com/b.json", "https://a.com/b.json"]
-  ] as [string, string][])(
+    ["TestFile", `${DEFAULT_BASE_URL}/bob/TestFile`, "bob"],
+    ["https://a.com/b.bob", "https://a.com/b.bob", "bob"],
+    ["TestFile", `${DEFAULT_BASE_URL}/json/TestFile`, "json"],
+    ["https://a.com/b.json", "https://a.com/b.json", "json"]
+  ] as [string, string, string][])(
     "fetches a file from the server",
     (
       inputFile: string,
       resolvedFile: string,
+      filetype: string,
       done: jest.DoneCallback
     ): void => {
       const mockSuccessResponse = {};
@@ -42,9 +45,8 @@ describe("<WidgetFromJson>", (): void => {
         .spyOn(global as GlobalFetch, "fetch")
         .mockImplementation((): Promise<{}> => mockFetchPromise);
 
-      // Just call the function, throw away the result
-      shallow(
-        <WidgetFromJson
+      const wrapper = shallow(
+        <EmbeddedDisplay
           containerStyling={{
             position: "relative",
             height: "",
@@ -56,6 +58,7 @@ describe("<WidgetFromJson>", (): void => {
             maxWidth: ""
           }}
           file={inputFile}
+          filetype={filetype}
         />
       );
 
@@ -63,17 +66,26 @@ describe("<WidgetFromJson>", (): void => {
       expect(globalWithFetch.fetch).toHaveBeenCalledWith(resolvedFile);
 
       process.nextTick((): void => {
+        // 6
+        expect(wrapper.type()).toEqual(Display);
+
         globalWithFetch.fetch.mockClear(); // 7
         done(); // 8
       });
     }
   );
-
-  it("fetches converts to JSON", (done): void => {
-    const mockSuccessResponse = { type: "display", position: "relative" };
-    const mockJsonPromise = Promise.resolve(mockSuccessResponse);
+  it("converts a simple widget", (done): void => {
+    const mockSuccessResponse = `
+    <widget type="label" version="2.0.0">
+        <name>Label</name>
+        <text>From .bob file</text>
+        <x>30</x>
+        <y>10</y>
+        <width>140</width>
+    </widget>`;
+    const mockTextPromise = Promise.resolve(mockSuccessResponse);
     const mockFetchPromise = Promise.resolve({
-      json: (): Promise<{}> => mockJsonPromise
+      text: (): Promise<{}> => mockTextPromise
     });
 
     // Hack to satisfy typescript
@@ -87,7 +99,7 @@ describe("<WidgetFromJson>", (): void => {
       .mockImplementation((): Promise<{}> => mockFetchPromise);
 
     const wrapper = shallow(
-      <WidgetFromJson
+      <EmbeddedDisplay
         containerStyling={{
           position: "relative",
           height: "",
@@ -99,32 +111,42 @@ describe("<WidgetFromJson>", (): void => {
           maxWidth: ""
         }}
         file="TestFile"
+        filetype="bob"
       />
     );
 
     expect(globalWithFetch.fetch).toHaveBeenCalledTimes(1);
     expect(globalWithFetch.fetch).toHaveBeenCalledWith(
-      `${DEFAULT_BASE_URL}/json/TestFile`
+      `${DEFAULT_BASE_URL}/bob/TestFile`
     );
 
     process.nextTick((): void => {
       // 6
-      expect(wrapper.type()).toEqual(Display);
+      expect(wrapper.type()).toEqual(Label);
 
       globalWithFetch.fetch.mockClear(); // 7
       done(); // 8
     });
   });
 
-  it("fetches converts children to JSON", (done): void => {
-    const mockSuccessResponse = {
-      type: "display",
-      position: "relative",
-      children: [{ type: "label", position: "relative", text: "Test" }]
-    };
-    const mockJsonPromise = Promise.resolve(mockSuccessResponse);
+  it("converts a display with child widget", (done): void => {
+    const mockSuccessResponse = `
+    <?xml version="1.0" encoding="UTF-8"?>
+    <display version="2.0.0">
+        <name>Display</name>
+        <width>200</width>
+        <height>350</height>
+        <widget type="label" version="2.0.0">
+            <name>Label</name>
+            <text>From .bob file</text>
+            <x>30</x>
+            <y>10</y>
+            <width>140</width>
+        </widget>
+    </display>`;
+    const mockTextPromise = Promise.resolve(mockSuccessResponse);
     const mockFetchPromise = Promise.resolve({
-      json: (): Promise<{}> => mockJsonPromise
+      text: (): Promise<{}> => mockTextPromise
     });
 
     // Hack to satisfy typescript
@@ -138,7 +160,7 @@ describe("<WidgetFromJson>", (): void => {
       .mockImplementation((): Promise<{}> => mockFetchPromise);
 
     const wrapper = shallow(
-      <WidgetFromJson
+      <EmbeddedDisplay
         containerStyling={{
           position: "relative",
           height: "",
@@ -150,6 +172,63 @@ describe("<WidgetFromJson>", (): void => {
           maxWidth: ""
         }}
         file="TestFile"
+        filetype="bob"
+      />
+    );
+
+    expect(globalWithFetch.fetch).toHaveBeenCalledTimes(1);
+    expect(globalWithFetch.fetch).toHaveBeenCalledWith(
+      `${DEFAULT_BASE_URL}/bob/TestFile`
+    );
+
+    process.nextTick((): void => {
+      // 6
+      expect(wrapper.type()).toEqual(Display);
+      expect(wrapper.childAt(0).type()).toEqual(Display);
+      expect(
+        wrapper
+          .childAt(0)
+          .childAt(0)
+          .type()
+      ).toEqual(Label);
+
+      globalWithFetch.fetch.mockClear(); // 7
+      done(); // 8
+    });
+  });
+
+  it("converts fetched children to JSON", (done): void => {
+    const mockSuccessResponse =
+      '{ "type": "display", "position": "relative", "children": [{ "type": "label", "position": "relative", "text": "Test" }] }';
+    const mockJsonPromise = Promise.resolve(mockSuccessResponse);
+    const mockFetchPromise = Promise.resolve({
+      text: (): Promise<{}> => mockJsonPromise
+    });
+
+    // Hack to satisfy typescript
+    interface GlobalFetch extends NodeJS.Global {
+      fetch: any;
+    }
+    const globalWithFetch = global as GlobalFetch;
+
+    jest
+      .spyOn(global as GlobalFetch, "fetch")
+      .mockImplementation((): Promise<{}> => mockFetchPromise);
+
+    const wrapper = shallow(
+      <EmbeddedDisplay
+        containerStyling={{
+          position: "relative",
+          height: "",
+          width: "",
+          margin: "",
+          padding: "",
+          border: "",
+          minWidth: "",
+          maxWidth: ""
+        }}
+        file="TestFile"
+        filetype="json"
       />
     );
 
@@ -161,7 +240,13 @@ describe("<WidgetFromJson>", (): void => {
     process.nextTick((): void => {
       // 6
       expect(wrapper.type()).toEqual(Display);
-      expect(wrapper.childAt(0).type()).toEqual(Label);
+      // Why the nesting?
+      expect(
+        wrapper
+          .childAt(0)
+          .childAt(0)
+          .type()
+      ).toEqual(Label);
 
       globalWithFetch.fetch.mockClear(); // 7
       done(); // 8
