+++ src/ui/widgets/Image/image.tsx
@@ -20,7 +20,6 @@ export const ImageComponent = (
   props: InferWidgetProps<typeof ImageProps> & Component
 ): JSX.Element => {
   const baseUrl = useContext(BaseUrlContext);
-  console.log(`base url ${baseUrl}`);
   let file = `img/${props.src}`;
   if (!file.startsWith("http")) {
     file = `${baseUrl}/${file}`;
