+++ src/ui/widgets/index.ts
@@ -1,6 +1,7 @@
 export { ActionButton } from "./ActionButton/actionButton";
 export { Display } from "./Display/display";
 export { DynamicPageWidget } from "./DynamicPage/dynamicPage";
+export { EmbeddedDisplay } from "./EmbeddedDisplay/embeddedDisplay";
 export { FlexContainer } from "./FlexContainer/flexContainer";
 export { GroupingContainer } from "./GroupingContainer/groupingContainer";
 export { Image } from "./Image/image";
@@ -11,5 +12,3 @@ export { ProgressBar } from "./ProgressBar/progressBar";
 export { Readback } from "./Readback/readback";
 export { Shape } from "./Shape/shape";
 export { SlideControl } from "./SlideControl/slideControl";
-export { WidgetFromBob } from "./FromBob/fromBob";
-export { WidgetFromJson } from "./FromJson/fromJson";
